+++ src/main/java/com/prgrms/rg/domain/ridingpost/application/impl/RidingPostServiceImpl.java
@@ -5,7 +5,7 @@ import static com.google.common.base.Preconditions.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.prgrms.rg.domain.ridingpost.model.RidingImageSaveManagement;
+import com.prgrms.rg.domain.ridingpost.model.image.RidingImageSaveManagement;
 import com.prgrms.rg.domain.ridingpost.model.RidingPost;
 import com.prgrms.rg.domain.ridingpost.model.RidingSaveManagement;
 import com.prgrms.rg.domain.ridingpost.application.RidingPostService;
