+++ src/main/java/com/prgrms/rg/domain/ridingpost/model/RidingPost.java
@@ -22,6 +22,7 @@ import com.prgrms.rg.domain.common.file.model.AttachedImage;
 import com.prgrms.rg.domain.common.file.model.ImageOwner;
 import com.prgrms.rg.domain.common.file.model.TemporaryImage;
 import com.prgrms.rg.domain.common.model.BaseTimeEntity;
+import com.prgrms.rg.domain.ridingpost.model.image.RidingThumbnailImage;
 import com.prgrms.rg.domain.user.model.User;
 
 import lombok.Builder;
